@@ -5,13 +5,9 @@ import { Component } from '@angular/core';
   template: `
 <div class="wrap">
     <div class="floatleft"><app-currency-converter>Loading Application...</app-currency-converter></div>
-     
-   
-    
+       
 </div>
   
-    
- 
   `
 })
 export class ConverterComponent {
